Broke down logic in oauth2callback into a helper function, verify_credentials_and_login.
This was done because it will actually make the oauth2 logic testable. It's impossible to automate the process to make valid credentials without storing a real account's password publically. It'll also be incompatible with fixture. A better solution is to use httpmocksequence to create a set of credentials. The only thing the routing for the oauth2callback should be doing is receiving the code and exchanging it.