Integrated Flask-Login to avoid global objects and session mangling.
This change removes requires_login usage and replaces it with login_manager.login_required. User Model has been updated to use the UserMixin object to provide required Flask-Login  functions. Templates have been updated to no longer user g.user.

This was done because it was impossible to log a user in properly for testing due to session mismatching.